# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
#host: localhost
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
produces:
  - application/json
paths:
  /censos:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getCensos
      parameters:
        - name: entidad 
          in: query
          type: string
          required: true
        - name: municipio
          in: query
          type: string
          required: true

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCensoListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /internet:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getInternetbyMun
      parameters:
        - name: entidad
          in: query
          type: string
          required: true
        - name: municipio
          in: query
          type: string
          required: true

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetInternetResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /idh:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getIdhbyMun
      parameters:
        - name: entidad
          in: query
          type: string
          required: true
        - name: municipio
          in: query
          type: string
          required: true

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIdhResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /denues:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getGameSystems
      parameters:
        - name: entidad
          in: query
          type: string
          required: true
        - name: municipio
          in: query
          type: string
          required: true

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetGameSystemListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /entidades:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getEntidades

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetEntidadListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /municipios:
    x-swagger-router-controller: analiticafin.compute.controller
    get:
      description: get the game system list
      operationId: getMunicipios
      parameters:
        - name: entidad
          in: query
          type: string
          required: true

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetMunicipioListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  GetInternetResponse:
    type: object
    properties:
      #id:
      #  type: string
      #  description: identifier
      entidad:
        type: string
        description: Name of the Game System
      municipio:
        type: string
        description: Description of the Game System
      habitadas:
        type: string
        description: gps latitud
      viviendas_con_computadora:
        type: string
        description: gps longitud
      viviendas_acceso_internet:
        type: string
        description: gps longitud
      cve_ent:
        type: string
        description: gps longitud
      cve_mun:
        type: string
        description: gps longitud
  GetIdhResponse:
    type: object
    properties:
      #id:
      #  type: string
      #  description: identifier
      entidad:
        type: string
        description: Name of the Game System
      municipio:
        type: string
        description: Description of the Game System
      promedio_escolaridad:
        type: string
        description: gps latitud
      escolaridad_esperada:
        type: string
        description: gps longitud
      ipc_usd:
        type: string
        description: gps longitud
      cve_ent:
        type: string
        description: gps longitud
      cve_mun:
        type: string
        description: gps longitud
      mortalidad_infantil:
        type: string
        description: gps longitud
      indice_educacion:
        type: string
        description: gps longitud
      indice_ingreso:
        type: string
        description: gps longitud
      indice_salud:
        type: string
        description: gps longitud
      idh:
        type: string
        description: gps longitud
...
  GetGameSystemResponse:
    type: object
    properties:
      id:
        type: string
        description: identifier
      nom_estab:
        type: string
        description: Name of the Game System
      nom_vial:
        type: string
        description: Description of the Game System
      latitud:
        type: string
        description: gps latitud
      longitud:
        type: string
        description: gps longitud
  GetGameSystemListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetGameSystemResponse"
  GetCensoResponse:
    type: object
    properties:
      #   id:
      #      type: string
      #      description: identifier
      entidad:
        type: string
        description: Name of the Game System
      municipio:
        type: string
        description: Description of the Game System
      actividad_economica:
        type: string
        description: gps latitud
      ue:
        type: string
        description: gps longitud
  GetCensoListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetCensoResponse"
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  #  GeneralResponse:
  #    type: object
  #    properties:
  #      success:
  #        type: number
  #        description: returns 1 if successful
  #      description:
  #        type: string
  #        description: a short comment
  #    required:
  #      - success
  #      - description
...
  GetMunicipioResponse:
    type: object
    properties:
      id:
        type: string
        description: identifier
      cve_mun:
        type: string
        description: Name of the game
      municipio:
        type: string
        description: Developer of the game
  GetMunicipioListResponse:
    required:
      - videogames
    properties:
      videogames:
        type: array
        items:
          $ref: "#/definitions/GetMunicipioResponse"
  GetEntidadResponse:
    type: object
    properties:
      id:
        type: string
        description: identifier
      cve_ent:
        type: string
        description: Name of the game
      entidad:
        type: string
        description: Developer of the game
  GetEntidadListResponse:
    required:
      - videogames
    properties:
      videogames:
        type: array
        items:
          $ref: "#/definitions/GetEntidadResponse"
...